Separated main and tests Precode removed
